pyproject.toml
[project]
name = "fetch-podcast"
version = "0.1.0"
description = "CLI utility to fetch podcast episodes from Spotify and export to CSV"
readme = "README.md"
requires-python = ">=3.12"
dependencies = ["click>=8.1.0", "python-dotenv>=1.0.0"]

[project.scripts]
fetch-podcast = "fetch_podcast.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/fetch_podcast"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[dependency-groups]
dev = ["pytest>=9.0.2", "pytest-mock>=3.15.1"]
